Errors fixed, new Method
If you went to fetch an association, for example, after insert into a
database, it would attempt to fetch the association and thrown an
exception. This is because a table was not selected, prior to fetching.
I've now implemented a check to see if a table was previously selected.
New method: insertInto().